requirement.txt
braintree==3.19.0
Django==1.8.3
django-crispy-forms==1.4.0
django-filter==0.11.0
djangorestframework==3.2.3
djangorestframework-jwt==1.7.2
httpie==0.9.2
Markdown==2.6.2
Pillow==2.9.0
Pygments==2.0.2
PyJWT==1.4.0
requests==2.7.0
wheel==0.24.0
